+++ src/design/plone/contenttypes/behaviors/info_testata.py
@@ -3,10 +3,12 @@ from design.plone.contenttypes import _
 from plone.app.textfield import RichText
 from plone.autoform.interfaces import IFormFieldProvider
 from plone.dexterity.interfaces import IDexterityContent
+from plone.namedfile import field as namedfile
 from plone.supermodel import model
 from zope import schema
 from zope.component import adapter
-from zope.interface import provider, implementer
+from zope.interface import implementer
+from zope.interface import provider
 
 
 @provider(IFormFieldProvider)
@@ -25,7 +27,14 @@ class IInfoTestata(model.Schema):
             default="Inserisci eventuale testo informativo che verrà mostrato in testata.",  # noqa
         ),
     )
-
+    immagine_testata = namedfile.NamedBlobImage(
+        title=_(
+            u"immagine_testata_label",
+            default=u"Immagine aggiuntiva per la testata",
+        ),
+        description=u"Inserisci un'eventuale immagine da mostrare in testata.",
+        required=False,
+    )
     ricerca_in_testata = schema.Bool(
         title=_(u"ricerca_in_testata_label", default=u"Ricerca in testata"),
         default=False,
@@ -49,7 +58,12 @@ class IInfoTestata(model.Schema):
     model.fieldset(
         "testata",
         label=_("testata_fieldset_label", default=u"Testata"),
-        fields=["info_testata", "ricerca_in_testata", "mostra_navigazione"],
+        fields=[
+            "info_testata",
+            "immagine_testata",
+            "ricerca_in_testata",
+            "mostra_navigazione",
+        ],
     )