+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 1.0.1 (unreleased)
 ------------------
 
+- Add `immagine_testata` new field in *design.plone.contenttypes.behavior.info_testata* behavior.
+  [cekk]
 - Add `correlato_in_evidenza` new field in *design.plone.contenttypes.behavior.argomenti* behavior.
   [cekk]